+++ Java/DormitoryNotebooksJava/src/main/java/kaczuch/master_thesis/service/UserService.java
@@ -7,6 +7,6 @@ import java.util.Optional;
 
 public interface UserService {
 	User save(UserDto userDto);
-	Optional<User> findById(Long id);
+	Optional<User> findById(Integer id);
 
 }
